requirements.txt
Django==4.0.5
PyYAML==6.0
tweepy==4.10.0
